Migrated registering over to redux toolkit. Moved features folder into redux. Removed old code used for login
