--- PlateUp-reactnative/redux/actionTypes.js
@@ -1,1 +1,0 @@
-export const USER_LOGGED_IN = "userLoggedIn";
